+++ packages/modlink_ui/modlink_ui/shared/inputs/__init__.py
@@ -1,3 +1,4 @@
+from .combo_box import remove_combo_popup_outer_margin
 from .token_line_edit import TokenLineEdit
 
-__all__ = ["TokenLineEdit"]
+__all__ = ["TokenLineEdit", "remove_combo_popup_outer_margin"]